Detección de incendios forestales con Python/requirements.txt
numpy==2.0.2
Pillow==11.1.0
matplotlib==3.10.0
scikit-learn==1.6.1
keras==3.8.0